+++ packages/compile/src/generate/gen-code.js
@@ -4,23 +4,21 @@ import { generateJS } from './gen-code-js.js'
 /**
  * Generate code from the given parsed model.
  *
- * @param {*} parsedModel The parsed model structure.
+ * @param {import('../parse-and-generate.js').ParsedModel} parsedModel The parsed model structure.
  * @param {Object} opts The options that control code generation.
- * @param {Object} opts.spec The parsed `spec.json` object.
- * @param {string[]} opts.operations The array of operations to perform.
- * - If it has 'generateC', the generated C code will be written to `buildDir`.
- * - If it has 'generateJS', the generated JS code will be written to `buildDir`.
- * - If it has 'printRefIdTest', reference identifiers will be printed to the console.
- * - If it has 'convertNames', no output will be generated, but the results of model
- *   analysis will be available.
- * @param {Map<string, any>} opts.extData The map of datasets from external `.dat` files.
- * @param {Map<string, any>} opts.directData The mapping of dataset name used in a
+ * @param {import('../_shared/model-spec.js').ModelSpec} opts.spec The parsed `spec.json` object.
+ * @param {import('../parse-and-generate.js').GenerateOperation[]} opts.operations The array
+ * of operations to perform.
+ * @param {import('../_shared/read-dat.js').ExtData} [opts.extData] The map of datasets from
+ * external `.dat` files.
+ * @param {Map<string, any>} [opts.directData] The mapping of dataset name used in a
  * `GET DIRECT DATA` call (e.g., `?data`) to the tabular data contained in the loaded
  * data file.
- * @param {string} opts.modelDirname The absolute path to the directory containing data
+ * @param {string} [opts.modelDirname] The absolute path to the directory containing data
  * (dat, xlsx, csv) files that are referenced by the model.  This path is used for
  * resolving data files for `GET DIRECT SUBSCRIPT` calls.
- * @returns A string containing the generated code.
+ * @param {string} [opts.varname] The variable name passed to the `sde causes` command.
+ * @returns {string} A string containing the generated code.
  */
 export function generateCode(parsedModel, opts) {
   // Note that the two `generate` functions perform the same steps (other than the